+++ packages/framework-integration-tests/integration/helper/expect.ts
@@ -1,7 +1,10 @@
 import * as chai from 'chai'
+import * as sinonChai from 'sinon-chai'
+import * as chaiAsPromised from 'chai-as-promised'
+import * as chaiArrays from 'chai-arrays'
 
-chai.use(require('sinon-chai'))
-chai.use(require('chai-as-promised'))
-chai.use(require('chai-arrays'))
+chai.use(sinonChai)
+chai.use(chaiAsPromised)
+chai.use(chaiArrays)
 
 export const expect = chai.expect