+++ packages/framework-integration-tests/integration/provider-specific/kubernetes/nuke/nuke.integration.ts
@@ -1,9 +1,7 @@
-import * as chai from 'chai'
 import { K8SManager } from '../k8s-manager'
 import { waitForIt } from '../../../helper/sleep'
 import { V1PodList, V1ServiceList } from '@kubernetes/client-node'
-
-export const expect = chai.expect
+import { expect } from '../../../helper/expect'
 
 describe('After nuke', () => {
   const k8sManager = new K8SManager()