+++ packages/framework-integration-tests/integration/provider-unaware/end-to-end/subscriptions.integration.ts
@@ -1,7 +1,6 @@
 import { random } from 'faker'
 import gql from 'graphql-tag'
-import { expect } from 'chai'
-import * as chai from 'chai'
+import { expect } from '../../helper/expect'
 import { Observable } from 'apollo-client/util/Observable'
 import { waitForIt } from '../../helper/sleep'
 import { FilterFor } from '@boostercloud/framework-types'
@@ -9,8 +8,6 @@ import { DisconnectableApolloClient } from '@boostercloud/application-tester'
 import { applicationUnderTest } from './setup'
 import { beforeHookProductId } from '../../../src/constants'
 import * as path from 'path'
-
-chai.use(require('chai-as-promised'))
 
 describe('subscriptions', () => {
   //TODO: Azure provider doesn't support subscription Interface so these tests are skipped for Azure