+++ packages/framework-provider-aws/src/setup.ts
@@ -13,6 +13,7 @@ import {
   readEntityEventsSince,
   readEntityLatestSnapshot,
   storeEvents,
+  storeSnapshot,
 } from './library/events-adapter'
 import { searchEntitiesIds, searchEvents } from './library/events-searcher-adapter'
 import { rawGraphQLRequestToEnvelope } from './library/graphql-adapter'
@@ -62,8 +63,9 @@ export const Provider = (rockets?: RocketDescriptor[]): ProviderLibrary => {
       forEntitySince: readEntityEventsSince.bind(null, dynamoDB),
       latestEntitySnapshot: readEntityLatestSnapshot.bind(null, dynamoDB),
       search: searchEvents.bind(null, dynamoDB),
-      store: storeEvents.bind(null, dynamoDB),
       searchEntitiesIDs: searchEntitiesIds.bind(null, dynamoDB),
+      store: storeEvents.bind(null, dynamoDB),
+      storeSnapshot: storeSnapshot.bind(null, dynamoDB),
     },
     // ProviderReadModelsLibrary
     readModels: {